Brain-Tumour-Image-classification-app-main/requirements.txt
streamlit
pandas
numpy
seaborn
tensorflow
Pillow
gdown
